+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  - Depends on Magic 1.8
  - Added tracking of wins/losses
  - Added leaderboard tracking
+ - Implement physical leaderboard
 
 ## 1.0